+++ packages/json-rpc/json-rpc-provider-proxy/src/get-sync-provider.ts
@@ -32,7 +32,7 @@ export const getSyncProvider =
               return cb(onMsg, (e) => {
                 if (isOn) {
                   isOn = false
-                  const diff = lastHalt - Date.now()
+                  const diff = Date.now() - lastHalt
                   consecutiveHalts +=
                     diff > WAIT_BASE + getWaitTime() ? -consecutiveHalts : 1
                   lastHalt += diff