+++ packages/tx-utils/src/decode-extrinsic.ts
@@ -10,7 +10,6 @@ import {
   StringRecord,
   Struct,
   u8,
-  UnifiedMetadata,
 } from "@polkadot-api/substrate-bindings"
 import { getMetadata } from "./get-metadata"
 import { getDynamicBuilder, getLookupFn } from "@polkadot-api/metadata-builders"
@@ -44,15 +43,17 @@ export const getExtrinsicDecoder = (
   const lookup = getLookupFn(metadata)
   const dynamicBuilder = getDynamicBuilder(lookup)
 
-  const innerExtra = Object.fromEntries(
-    metadata.extrinsic.signedExtensions.map(
-      (x) =>
-        [x.identifier, dynamicBuilder.buildDefinition(x.type)[1]] as [
-          string,
-          Decoder<any>,
-        ],
-    ),
-  ) as StringRecord<Decoder<any>>
+  const v0Extra = Struct.dec(
+    Object.fromEntries(
+      metadata.extrinsic.signedExtensions[0].map(
+        (x) =>
+          [x.identifier, dynamicBuilder.buildDefinition(x.type)[1]] as [
+            string,
+            Decoder<any>,
+          ],
+      ),
+    ) as StringRecord<Decoder<any>>,
+  )
 
   let call: Decoder<any>
   let address: Decoder<any>
@@ -79,7 +80,7 @@ export const getExtrinsicDecoder = (
   const v4Body = Struct.dec({
     address,
     signature,
-    extra: Struct.dec(innerExtra),
+    extra: v0Extra,
     callData: allBytesDec,
   })
 
@@ -94,20 +95,22 @@ export const getExtrinsicDecoder = (
       const extensionVersion = u8.dec(data)
       let extraDec: Decoder<StringRecord<any>>
       if (metadata.version === 16) {
-        const extensionsToApply = (
-          metadata as UnifiedMetadata<16>
-        ).extrinsic.signedExtensionsByVersion.find(
-          ([x]) => x === extensionVersion,
-        )
+        const extensionsToApply =
+          metadata.extrinsic.signedExtensions[extensionVersion]
+
         if (!extensionsToApply) throw new Error("Unexpected extension version")
         extraDec = Struct.dec(
           Object.fromEntries(
-            Object.entries(innerExtra).filter((_, idx) =>
-              extensionsToApply[1].includes(idx),
+            extensionsToApply.map(
+              (x) =>
+                [x.identifier, dynamicBuilder.buildDefinition(x.type)[1]] as [
+                  string,
+                  Decoder<any>,
+                ],
             ),
           ) as StringRecord<Decoder<any>>,
         )
-      } else extraDec = Struct.dec(innerExtra)
+      } else extraDec = v0Extra
       const extra = extraDec(data)
 
       return {