+++ packages/substrate-bindings/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+### Changed
+
+- `UnifiedMetadata` now exposes signedExtensions by `ExtVersion`
+
+## 0.16.6 - 2026-01-05
+
+### Fixed
+
+- Binary: opaque representation
+
 ## 0.16.5 - 2025-10-31
 
 ### Fixed